requirements.txt
line-bot-sdk
openai
requests
python-dotenv
azure-cognitiveservices-speech
Flask
azure-storage-blob
